fix: 엔티티 타입 변경

+++ src/main/java/com/mogak/spring/repository/MogakCategoryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MogakCategoryRepository extends JpaRepository<MogakCategory, Long> {
+public interface MogakCategoryRepository extends JpaRepository<MogakCategory, Integer> {
     Optional<MogakCategory> findMogakCategoryByName(String name);
 }
